+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3975,7 +3975,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KWE79M6HFSS3MFXBGYQC3F54"),
+  id: z.string().default("01KWECGYE968WX5APP0M6QE84D"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),